% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cpython_paths.R
\name{cpython_interpreter_path}
\alias{cpython_interpreter_path}
\title{Path to the bundled CPython interpreter}
\usage{
cpython_interpreter_path(must_exist = TRUE)
}
\arguments{
\item{must_exist}{Logical. If \code{TRUE} (default), error when no interpreter
is found.}
}
\value{
A single string path to the interpreter.
}
\description{
Returns the on-disk path to the CPython interpreter shipped with the
package. If a built interpreter is not present, you can set
\code{must_exist = FALSE} to get the expected default location.
}
