% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPython_version.R
\name{cpython_version}
\alias{cpython_version}
\title{CPython version}
\usage{
cpython_version()
}
\value{
A single string version (for example, "3.14.2").
}
\description{
Reads the bundled CPython version from the embedded CPython runtime.
}
